move utility functions from azp.py to util.py
