+++ region/azp.py
@@ -4,8 +4,9 @@ import libpysal as ps
 from geopandas import GeoDataFrame
 import networkx as nx
 
-from region.util import dataframe_to_dict, dissim_measure, \
-                        find_sublist_containing
+from region.util import dataframe_to_dict, \
+                        distribute_regions_among_components,\
+                        find_sublist_containing, make_move, objective_func
 
 
 def azp(areas, data, num_regions, contiguity=None, initial_sol=None):
@@ -183,60 +184,3 @@ def _azp_connected_component(graph, initial_clustering, num_regions):
 
         obj_val_end = objective_func(region_list_copy, graph)
     return region_list_copy
-
-
-def distribute_regions_among_components(num_regions, nx_graph):
-    """
-
-    Parameters
-    ----------
-    num_regions : `int`
-        The overall number of regions.
-    nx_graph : `networkx.Graph`
-        An undirected graph whose number of connected components is not greater
-        than `num_regions`.
-
-    Returns
-    -------
-    result : `dict`
-        Each key (of type `networkx.Graph`) is a connected component of
-        `nx_graph`.
-        Each value is an `int` defining the number of regions in the key
-        component.
-    """
-    comps = list(nx.connected_component_subgraphs(nx_graph, copy=False))
-    num_regions_to_distribute = num_regions
-    result = {}
-    comps_multiplied = []
-    # make sure each connected component has at least one region assigned to it
-    for comp in comps:
-        comps_multiplied += [comp] * (len(comp)-1)
-        result[comp] = 1
-        num_regions_to_distribute -= 1
-    # distribute the rest of the regions to random components with bigger
-    # components being likely to get more regions assigned to them
-    while num_regions_to_distribute > 0:
-        position = random.randrange(len(comps_multiplied))
-        picked_comp = comps_multiplied.pop(position)
-        result[picked_comp] += 1
-        num_regions_to_distribute -= 1
-    return result
-
-
-def objective_func(region_list, graph, attr="data"):
-    return sum(dissim_measure(graph.node[list(region_list[r])[i]][attr],
-                              graph.node[list(region_list[r])[j]][attr])
-               for r in range(len(region_list))
-               for i in range(len(region_list[r]))
-               for j in range(len(region_list[r]))
-               if i < j)
-
-
-def make_move(area, from_region, to_region, region_list_copy):
-    print("  move", area, "from", from_region, "to", to_region)
-    from_index = region_list_copy.index(from_region)
-    to_index = region_list_copy.index(to_region)
-    from_region.remove(area)
-    to_region.add(area)
-    region_list_copy[from_index] = from_region
-    region_list_copy[to_index] = to_region
